再生成時のドキュメント状態を修正
1. **`applyChatDiff` / `getChatFromCache` の共通化** ([`app/lib/chatHistory.ts`](file:///Users/kou/projects/my-code/app/lib/chatHistory.ts))
   - `applyChatDiff` および `getChatFromCache` を共通ライブラリ `chatHistory.ts` に配置・export し、Server Component から安全に再利用できるように整備しました。

2. **対象チャットより前の diff のみを適用したドキュメント状態の算出**
   - チャットページ表示時（Server Component）にて、対象ページの全チャット履歴を取得。
   - 再生成対象のチャットの作成日時 (`createdAt`) よりも**古い**チャットの履歴のみを抽出 (`priorChatHistories`)。
   - `applyChatDiff(sections, priorChatHistories)` を実行し、**「そのチャットが作成された時点のドキュメント内容」**（＝対象チャット以降の diff は適用されていない過去状態）を正確に復元した `priorSectionContent` を計算・生成しました。

3. **再生成リクエスト時のプロンプト適用**
   - 「再生成」ボタン押下時に、復元された `priorSectionContent` をそのまま `/api/chat` へ送信するように更新しました。

これにより、チャット再生成時に「対象チャットより前の時刻に作られた diff だけが適用された正確なドキュメントテキスト」に基づいて AI 回答・差分生成が行われます。
