+++ app/lib/chatHistory.ts
@@ -3,9 +3,18 @@ import { getAuthServer } from "./auth";
 import { getDrizzle } from "./drizzle";
 import { chat, diff, message, section } from "@/schema/chat";
 import { and, asc, eq, exists } from "drizzle-orm";
-import { updateTag } from "next/cache";
+import { cacheLife, cacheTag, updateTag } from "next/cache";
 import { isCloudflare } from "./detectCloudflare";
-import { LangId, PagePath, PageSlug, SectionId } from "./docs";
+import {
+  LangId,
+  MarkdownSection,
+  PagePath,
+  PageSlug,
+  ReplacedRange,
+  SectionId,
+  SectionWithDiff,
+} from "./docs";
+import { dateReviver } from "./dateReviver";
 
 export interface CreateChatMessage {
   role: "user" | "ai" | "error";
@@ -278,4 +287,103 @@ export async function migrateChatUser(oldUserId: string, newUserId: string) {
     .update(chat)
     .set({ userId: newUserId })
     .where(eq(chat.userId, oldUserId));
+}
+
+export function applyChatDiff(
+  splitMdContent: MarkdownSection[],
+  chatHistories: ChatWithMessages[]
+): SectionWithDiff[] {
+  const newContent: SectionWithDiff[] = splitMdContent.map((section) => ({
+    ...section,
+    replacedContent: section.rawContent,
+    replacedRange: [] as ReplacedRange[],
+  }));
+  const chatDiffs = chatHistories.map((chat) => chat.diff).flat();
+  chatDiffs.sort(
+    (a, b) => new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime()
+  );
+  for (const diff of chatDiffs) {
+    const targetSection = newContent.find((s) => s.id === diff.sectionId);
+    if (targetSection) {
+      const startIndex = targetSection.replacedContent.indexOf(diff.search);
+      if (startIndex !== -1) {
+        const endIndex = startIndex + diff.search.length;
+        const replaceLen = diff.replace.length;
+        const diffLen = replaceLen - diff.search.length; // 文字列長の増減分
+
+        // 1. 文字列の置換
+        targetSection.replacedContent =
+          targetSection.replacedContent.slice(0, startIndex) +
+          diff.replace +
+          targetSection.replacedContent.slice(endIndex);
+
+        // 2. 既存のハイライト範囲のズレを補正（今回の置換箇所より後ろにあるものをシフト）
+        targetSection.replacedRange = targetSection.replacedRange.map((h) => {
+          if (h.start >= endIndex) {
+            // 完全に後ろにある場合は単純にシフト
+            return {
+              start: h.start + diffLen,
+              end: h.end + diffLen,
+              id: h.id,
+            };
+          }
+          if (h.end >= endIndex) {
+            return { start: h.start, end: h.end + diffLen, id: h.id };
+          }
+          return h;
+        });
+
+        // 3. 今回の置換箇所を新たなハイライト範囲として追加
+        targetSection.replacedRange.push({
+          start: startIndex,
+          end: startIndex + replaceLen,
+          id: diff.chatId,
+        });
+      } else {
+        // TODO: md5ハッシュを参照し過去バージョンのドキュメントへ適用を試みる
+        console.error(
+          `Failed to apply diff: search string "${diff.search}" not found in section ${targetSection.id}`
+        );
+      }
+    } else {
+      console.error(
+        `Failed to apply diff: section with id "${diff.sectionId}" not found`
+      );
+    }
+  }
+
+  return newContent;
+}
+
+/**
+ * チャットの取得をキャッシュする。
+ *
+ * use cacheの仕様で、drizzleオブジェクトとauthオブジェクトは引数に渡せない。
+ * 一方、use cacheの関数内でheaders()にはアクセスできない。
+ * したがって、外でheaders()を使ってuserIdを取得した後、関数の中で再度drizzleを初期化しないといけない。
+ *
+ * docsとchatの2箇所のサーバーコンポーネントで使用。ServerActionやrouteではこれではなく直接getAllChat()を呼んだ方が確実なはずです。
+ */
+export async function getChatFromCache(path: PagePath, userId?: string) {
+  "use cache";
+  cacheLife("days");
+
+  if (!userId) {
+    return [];
+  }
+  cacheTag(cacheKeyForPage(path, userId));
+
+  if (isCloudflare()) {
+    const cache = await caches.open("chatHistory");
+    const cachedResponse = await cache.match(cacheKeyForPage(path, userId));
+    if (cachedResponse) {
+      const data = JSON.parse(
+        await cachedResponse.text(),
+        dateReviver
+      ) as ChatWithMessages[];
+      return data;
+    }
+  }
+  const ctx = await initContext({ userId });
+  return await getAllChat(path, ctx);
 }
